fix: only displayed during installation
